requirements.txt
PyQt5==5.15.10
keyboard==0.13.5
pyperclip==1.8.2
requests==2.31.0
python-dotenv==1.0.0
pyinstaller==6.3.0
pillow==10.2.0
googletrans==4.0.0rc1
psutil==5.9.8 